% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avalanchr-package.R
\docType{package}
\name{avalanchr-package}
\alias{avalanchr}
\alias{avalanchr-package}
\title{avalanchr: Tools for the Avalanche Data Science Team}
\description{
A set of tools to use within the Avalance Data
    Science team. This package contains functions for connecting to data
    bases, R markdown templates, ggplot2 themes, and more.
}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolm@avalanche.com}

Authors:
\itemize{
  \item Barret Wallace
}

Other contributors:
\itemize{
  \item AVALANCHE [copyright holder]
}

}
\keyword{internal}
