% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{a data dictionary for AVALANCHE DBs}
\format{A \code{tibble}}
\usage{
data_dictionary
}
\description{
a data dictionary for AVALANCHE DBs
}
\keyword{datasets}
