% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{gt_data_dictionary}
\alias{gt_data_dictionary}
\title{Create gt table for the data dictionary}
\usage{
gt_data_dictionary()
}
\value{
a \code{gt} table
}
\description{
Create gt table for the data dictionary
}
