% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{gt_donations}
\alias{gt_donations}
\title{Create gt table for donations by sector}
\usage{
gt_donations(donations = get_donation_data())
}
\arguments{
\item{donations}{a data frame created by \code{\link[=get_donation_data]{get_donation_data()}}}
}
\value{
a \code{gt} table
}
\description{
Create gt table for donations by sector
}
