% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_donations}
\alias{plot_donations}
\title{Plot donation counts by sector}
\usage{
plot_donations(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=count_donations]{count_donations()}}}
}
\value{
a ggplot
}
\description{
Plot donation counts by sector
}
\examples{

plot_donations()

}
